+++ src/People.tsx
@@ -1,0 +1,3 @@
+export interface Names {
+  names: string[];
+}
